Refactor for scalability
